Only build the slang library with 'static'
The default target cmake sets is 'install'. While a good approach in general,
the default slang install target comes with "everything" in it, implying
lengthier and more fragile builds. It also seems slang does not define
an install component that would be anywhere close to limited only to
the slang library. And the Rust cmake abstraction to not allow selecting
install components.

Given above, only build the library itself and take care set up linkage to Rust
from the *build* folder directly, instead of the install folder.